+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 0.3 (unreleased)
 ================
 
+- Support parallel reads in `open_ncml` using `dask`. By @huard
 - Closing the dataset returned by `open_ncml` will close the underlying opened files. By @huard
 - Add `add_aggregation` and `add_variable_agg` to `Dataset` class. By @huard
 - Add `add_scan` to `Dataset` class. By @huard
-
 
 
 0.2 (2023-02-23)